pillow/Image/Image/save/pyproject.toml
[tool.poetry]
name = "basic-example"
version = "0.1.0"
description = ""
authors = ["jongwoo Yoo <dbwhddn10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^6.2.5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
